+++ testing.R
@@ -3,5 +3,6 @@ library(survival)
 
 KM <- survfit(Surv(rectime, censrec) ~ hormon, data = brcancer)
 time_scale <- seq(0, max(brcancer$rectime), by = 365)
-p <- KMunicate(fit = KM, time_scale = time_scale, .reverse = TRUE, .risk_table_base_size = 6, .rel_heights = c(6, 1, 1))
+p <- KMunicate(fit = KM, time_scale = time_scale, .reverse = TRUE, .risk_table_base_size = 6, .rel_heights = c(6, 1, 1), .annotate = ggplot2::annotate("text", x = 1, y = 1, label = "Some text"))
+
 ggsave(p, filename = "test.png", height = 7, width = 7, dpi = 300)